cookmate-flutter-recommendation/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
flask>=2.0.0
python-dotenv>=0.19.0
pymongo>=4.0.0
requests>=2.25.0
flask-cors>=3.0.10